Add files via upload
This is a coding problem I did for a phone technical test. The goal here is print these IDs one at a time and to ensure that an ID doesn't get printed multiple times in a row unless it's the last ID. An Android ID always starts with an 'A' and is four characters long. An IPhone ID always starts with 'I' and is three characters long. At the end of each ID is the number of times the ID appears. This is always a single digit.

Unfortunately, I wasn't able to finish the problem, but I did come up with a game plan. Not wanting to give up just yet, I decided to keep on chugging and solve the problem on my own time using the game plan I made during the interview.